Time Class Refactoring
Still need to implement query conversions and a fixed timestep for physics calculations, but basic timekeeping works without issue.

+++ Framework/Utils/Random.h
@@ -12,7 +12,7 @@ struct RNG {
 inline void RNG::Seed(float time)
 {
     //Seed the RNG to the input time. 
-    srand(time);
+    srand((unsigned int)time);
 }
 
 inline int RNG::RandInt(int max)
